Add files via upload
Given a number ‘n’ and a position ‘p’, write an algorithm and define a function using ‘C’ program to check if the ‘p-th’ digit from left of ‘n’ is odd or even. For example, if ‘n’ is 3145782 and p is 4 then you have to check if 5 is odd or even. Since it is odd print ‘Odd’. Make your code accept numbers of larger size.