Add files via upload
Given an array of integers 'arr' as an input. Write a program that will print all the top elements in the array. An element is said to be top if it is greater than all the elements to its right side. And the rightmost element is always a top element.