Add files via upload
Given a list of registration numbers that are unordered for a class, find a registration number 'x' from the list of numbers. If 'x' is present in the list print 'Yes' else print 'No'. Also print the corresponding position of 'x' in sorted list, if 'x' is found and print -1 if 'x' is not found.