Add files via upload
Consider a set of numbers in a list data structure, insert a new element 'x' based on the input position. If position is '0', insert a new element 'x' at first position. If position is '3', insert a new element 'x' at index 3. If the position entered is an invalid index your program has to print -1.