Add files via upload
Design a queue Q1 that can accommodate both odd and even numbers in it. Create a menu-driven C program to enqueue, dequeue, and display the elements of Q1. While dequing, check whether the deleted element is odd, if so, place all odd elements in the odd array else, place all even numbers in the even array. Based on the user's choice perform the corresponding enqueue, dequeue, and display operation. After every operation, get consent from the user whether to perform any other option or not. If the user wants to continue he/she has to press1, so that the program provides another attempt to the user to perform a different operation. If the user presses other than 1 then, display the present state of the queue and also the content of odd and even array respectively. If odd/even arrays are empty then print -1 to indicate zero elements in the odd array and print -99 to indicate zero elements in the even array.